% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NZ_AreaUnits}
\alias{NZ_AreaUnits}
\title{NZ_AreaUnits}
\format{A data frame with 5 variables:
\describe{
\item{\code{year}}{the year of the study}
\item{\code{CAU_num}}{numeric Area Unit code}
\item{\code{CAU_name}}{name of Area Unit}
\item{\code{Indicator}}{kind of indicator: either "CAU_avg_NZD", "CAU_avg_NZD_score", "AUScore" or "AUScale"}
\item{\code{Value}}{value of the indicator, as a numerical}
}

For further details, see \url{https://www.otago.ac.nz/wellington/departments/publichealth/research/hirp/otago020194.html}}
\usage{
NZ_AreaUnits
}
\description{
New Zealand Indexes of Deprivation
}
\keyword{datasets}
